% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imCoords.R
\name{bary2cart}
\alias{bary2cart}
\alias{bary2cart3d}
\title{Conversion from bary to cartesian}
\usage{
bary2cart(x)

bary2cart3d(x)
}
\arguments{
\item{x}{point coordinate information}
}
\description{
Conversion from barycentric coordinates to cartesian coordinates
}

