% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imPois.R
\name{dcpm}
\alias{dcpm}
\alias{pcpm}
\title{BL Probability Distribution}
\usage{
dcpm(x, pars, ztrunc)

pcpm(q, pars, ztrunc)
}
\arguments{
\item{x}{quantiles}

\item{pars}{a numeric vector of parameters}

\item{ztrunc}{a logical value incidcating whether standard Poisson is truncated at 0.}

\item{q}{quantiles}
}
\description{
Density and distribution function for the Bickis and Lee's distribution with three parameters \eqn{xi_2}, \eqn{xi_1}, and \eqn{xi_0}.
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}
}

