% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imPois.R
\name{evfn}
\alias{evfn}
\title{Expected Value of Canonical Variable}
\usage{
evfn(y, pars, ztrunc)
}
\arguments{
\item{y}{a vector of observations}

\item{pars}{a numeric vector of parameters}

\item{ztrunc}{a logical value incidcating whether standard Poisson is truncated at 0.}
}
\description{
Expected value of canonical variable is computed using \code{integrate} function.
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}
}

