% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imPois.R, R/ztrunc.R
\name{evfn}
\alias{evfn}
\alias{evfn.ztrunc}
\alias{lapprox.ztrunc}
\alias{mh.ztrunc}
\title{Expected Value of Canonical Variable}
\usage{
evfn(y, pars)

evfn.ztrunc(y, pars)

mh.ztrunc(y, pars, len.chain = 10000, len.burnin = 1000)

lapprox.ztrunc(y, pars, const)
}
\arguments{
\item{y}{a vector of observations}

\item{pars}{a numeric vector of parameters}

\item{len.chain}{a length of MH chain}

\item{len.burnin}{a length of burn-in period}

\item{const}{adjustment constant}
}
\description{
Expected value of canonical variable is computed using \code{integrate} function.
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}
}

