% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{fn.evfn}
\alias{fn.evfn}
\alias{gr.evfn}
\title{Hyperparamter Optimization}
\usage{
fn.evfn(x)

gr.evfn(x)
}
\arguments{
\item{x}{parameter needs to be optimized}
}
\description{
Score and graident functions are defined
}

