% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optim.R
\name{fn.evfn}
\alias{fn.evfn}
\alias{gr.evfn}
\title{Objective And Gradient Vector Needed For Optimization}
\usage{
fn.evfn(x)

gr.evfn(x)
}
\arguments{
\item{x}{parameter needs to be optimized}
}
\description{
Objective And Gradient Vector Needed For Optimization
}

