% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imPois.R
\docType{package}
\name{imPois}
\alias{imPois}
\alias{imPois-package}
\title{Imprecise Inferential Framework for Poisson Sampling Model}
\description{
Imprecise probabilities introduced by Peter Walley in 1991 is the basis of this imprecise inferential framework.  The package provides a collection of tools for conducting analysis of epistemic uncerstainty with Poisson and zero-truncated Poisson sampling parameter estimation.
}

