% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imPois.R, R/ztrunc.R
\name{dcpm}
\alias{dcpm}
\alias{dcpm.ztrunc}
\alias{pcpm}
\alias{pcpm.ztrunc}
\title{Imprecise Probability Distribution}
\usage{
dcpm(x, pars)

pcpm(q, pars)

dcpm.ztrunc(m, pars, ny)

pcpm.ztrunc(q, pars, ny)
}
\arguments{
\item{x}{quantiles}

\item{pars}{a numeric vector of parameters}

\item{q}{quantiles}

\item{m}{random variable}

\item{ny}{number of observations}
}
\description{
Density and distribution function for the Bickis and Lee's distribution with three parameters \eqn{xi_2}, \eqn{xi_1}, and \eqn{xi_0}.
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}
}

