% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imPois.R
\docType{package}
\name{ipeglim}
\alias{ipeglim}
\alias{ipeglim-package}
\title{Imprecise Inferential Framework for GLIM}
\description{
The package 'ipeglim' (Imprecise Probability Estimates for GLIM)
is a collection of tools for conducting analysis of epistemic
uncertainty in the generalized linear model setup based on Walley's
(1991) Imprecise Probability Theory.
}

