% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imPois.R, R/ztrunc.R
\name{cgf}
\alias{cgf}
\alias{cgf.ztrunc}
\title{Comupting Normalizing Constant of Bickis and Lee's Probability Distribution}
\usage{
cgf(xi2, xi1, xi0, log = TRUE)

cgf.ztrunc(xi2, xi1, xi0, ny)
}
\arguments{
\item{xi2}{parameter associated with precision}

\item{xi1}{parameter associated with linear combination}

\item{xi0}{effective sample size}

\item{log}{a logical value; if TRUE, probabilities are given as \eqn{log(p)}.}

\item{ny}{number of observations}
}
\description{
Given parameters, a normalizing constant for Bickis and Lee's probability distribution is computed.
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}
}

