% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imGraphics.R
\name{plot.impinf}
\alias{pbox}
\alias{plot.impinf}
\title{Plotting Imprecise Objects}
\usage{
\method{plot}{impinf}(x, after.obs = FALSE, ...)

pbox(x, xmin = -4, xmax = 4, ncdfs = 5, x.by = 0.05, xmar, ...)
}
\arguments{
\item{x}{an object for which a plot is needed.}

\item{after.obs}{logical indicating imprecise prior or posterior.}

\item{...}{additional arguments affecting the plot produced.}

\item{xmin}{lower limit of quantile}

\item{xmax}{upper limit of quantile}

\item{ncdfs}{number of CDFs}

\item{x.by}{argument used in seq on cdf}

\item{xmar}{adjustment for label}
}
\description{
Generic function for plotting of \code{imprecise} objects

Generating a probability box
}

