% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.impinf}
\alias{pbox}
\alias{plot.impinf}
\title{Plotting Imprecise Objects}
\usage{
\method{plot}{impinf}(x, after.obs = FALSE, ...)

pbox(x, min, max, ...)
}
\arguments{
\item{x}{an object for which a plot is needed.}

\item{after.obs}{logical indicating imprecise prior or posterior.}

\item{...}{additional arguments affecting the plot produced.}

\item{min}{lower limit of a distribution}

\item{max}{upper limit of a distribution}
}
\description{
Generic function for plotting \code{imprecise} objects.

Constructing Probability Band
}

