% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imTools.R
\name{print.summary.impinf}
\alias{print.summary.impinf}
\title{Print Summarized Objects}
\usage{
\method{print}{summary.impinf}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\description{
Printing \code{impinf} objects
}

