% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{print.summary.impinf}
\alias{print.summary.impinf}
\title{Print Imprecise Objects}
\usage{
\method{print}{summary.impinf}(x, ...)
}
\arguments{
\item{x}{an object used to select a method}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print Imprecise Objects
}

