% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imTools.R
\name{summary.impinf}
\alias{summary.impinf}
\title{Summaries of \code{impinf} object}
\usage{
\method{summary}{impinf}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is needed.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summarizing outputs produced from \code{update.impinf}.
}

