% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{summary.impinf}
\alias{summary.impinf}
\title{Summary of \code{impinf} object}
\usage{
\method{summary}{impinf}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is needed.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary of \code{impinf} object
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}
}

