% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{update.impinf}
\alias{update.impinf}
\title{Applying Bayes Rule}
\usage{
\method{update}{impinf}(object, y, ...)
}
\arguments{
\item{object}{an object for which the Bayes rule is needed.}

\item{y}{a vector of observations}

\item{...}{further arguments passed to methods}
}
\description{
Applying Bayes Rule
}
\author{
Chel Hee Lee \email{chl948@mail.usask.ca}
}

