% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{iprior}
\alias{iprior}
\title{Characterize Imprecise Prior}
\usage{
iprior(ui, ci)
}
\arguments{
\item{ui}{constraint matrix (k x p), see below.}

\item{ci}{constrain vector of length k, see below.}
}
\description{
A set of linear inequalities is used for characterizing a polygonal convex set.  The \code{chull} function in the \code{grDevices} package and the \code{convhulln} function in the \code{geometry} package are used to search for extreme points constructing a convex hull.
}
\examples{
# lc0 <- list(lhs=rbind(diag(2), -diag(2)), rhs=c(0,0,-1,-1))
# op <- iprior(ui=rbind(diag(2), -diag(2)), ci=c(0,0,-1,-1))
# op <- iprior(ui=rbind(c(1,0),c(0,1),c(-1,-1)), ci=c(0,0,-5))
op <- iprior(ui=rbind(c(1,0),c(0,1),c(0,-1),c(1,1),c(-2,-1)),
   ci=c(1,2,-8,5,-14)) # (3,8),(1,8), (1,4),(3,2)(6,2)
}
\author{
Chel Hee Lee <gnustats@gmail.com>
}

